progress in the configurator
